Clean up join-point reload optimization
